+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Implement Campusonline Public REST API, caching the results in a local database table and getting organizations form there
+
 ## v0.1.22
 
 - Modernize code